vhpidirect/quickstart/wrapping/exitcb: cleanup after abort fix in ghdl/ghdl
